qc/rseqc: relocate to 'qc.apps.rseqc'.
